SDP/Module32_Final Exam/Household_Backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Brotli==1.1.0
click==8.1.7
dj-database-url==2.2.0
Django==4.2.15
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-phone-field==1.8.1
djangorestframework==3.14.0
gunicorn==23.0.0
h11==0.14.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pytz==2024.2
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
uvicorn==0.30.6
whitenoise==6.7.0
